+++ pages/studentAssessmentInstanceHomework/studentAssessmentInstanceHomework.sql
@@ -16,7 +16,11 @@ SELECT
     aq.max_points,
     qo.row_order,
     qo.question_number,
-    aq.init_points
+    aq.init_points,
+    z.max_points AS zone_max_points,
+    (z.max_points IS NOT NULL) AS zone_has_max_points,
+    z.best_questions AS zone_best_questions,
+    (z.best_questions IS NOT NULL) AS zone_has_best_questions
 FROM
     instance_questions AS iq
     JOIN assessment_questions AS aq ON (aq.id = iq.assessment_question_id)